+++ source/lantana/render/mesh/generic.d
@@ -244,7 +244,7 @@ template GenericMesh(Attrib, Loader, GlobalUniforms=DefaultUniforms, Settings = 
 					}
 				}
 
-				foreach(ulong i; 0..inst.mesh.bones.length)
+				foreach(size_t i; 0..inst.mesh.bones.length)
 				{
 					inst.anim.boneMatrices[i] = 
 						applyParentTransform(inst.anim.bones[i], inst.anim.bones)
